pyproject.toml
[tool.poetry]
name = "sentinel"
version = "0.2.6"
description = ""
authors = ["Vipul Sharma <vipul.sharma20@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.3.1"
boto3 = "^1.18.44"
requests = "^2.26.0"
PyYAML = "^5.4.1"
docopt = "^0.6.2"
tabulate = "^0.8.9"
pytest-cov = "^2.12.1"
slack-sdk = "^3.11.2"
togpush = {url = "http://cheeseshop.vernacular.ai/togpush/togpush-1.1.15-py3-none-any.whl"}
fsm-pull = "^0.2.29"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
sentinel = "sentinel.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
